ata drive writes + code cleanup here and there

+++ src/arch/x86_64/interrupts/exceptions.rs
@@ -116,12 +116,12 @@ exception_function!(0xFF, generic_handler, true);
 
 pub fn set_exceptions() {
     for i in 0..32 {
-        idt_set_gate(i, generic_handler as u64);
+        idt_set_gate(i, generic_handler as usize);
     }
 
-    idt_set_gate(0x00, div_error as u64);
-    idt_set_gate(0x06, invalid_opcode as u64);
-    idt_set_gate(0x08, double_fault as u64);
-    idt_set_gate(0x0D, general_protection_fault as u64);
-    idt_set_gate(0x0E, page_fault as u64);
+    idt_set_gate(0x00, div_error as usize);
+    idt_set_gate(0x06, invalid_opcode as usize);
+    idt_set_gate(0x08, double_fault as usize);
+    idt_set_gate(0x0D, general_protection_fault as usize);
+    idt_set_gate(0x0E, page_fault as usize);
 }
